mgr/cephadm: move _deploy_etc_ceph_ceph_conf to serve.py
